pyproject.toml
[tool.poetry]
name = "ohtugroup2"
version = "0.1.0"
description = ""
authors = ["Alexander, Joonas, Meeri, Neoh, Noora, Juuso"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
coverage = "^6.5.0"
pytest = "^7.2.0"
attr = "^0.3.2"
gunicorn = "^20.1.0"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
selenium = "^4.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
